requirements.txt

matplotlib >= 3.2.1
geopandas >= 0.7.0
descartes >= 1.1.0
